move code to tokens.ts
