+++ fuse-subgraph/subgraph.template.yaml
@@ -75,7 +75,7 @@ templates:
       kind: ethereum/events
       apiVersion: 0.0.3
       language: wasm/assemblyscript
-      file: ./src/mappings/bridge.ts
+      file: ./src/mappings/tokens.ts
       entities:
         - Token
         - TransferEvent
